test in-house estimator separately
