+++ packages/datadog-plugin-openai/test/token-estimator.spec.js
@@ -1,0 +1,28 @@
+'use strict'
+
+const { estimateTokens } = require('../src/token-estimator')
+
+describe('Plugin', () => {
+  describe('openai token estimation', () => {
+    function testEstimation (input, expected) {
+      const tokens = estimateTokens(input)
+      expect(tokens).to.equal(expected)
+    }
+
+    it('should compute the number of tokens in a string', () => {
+      testEstimation('hello world', 2)
+    })
+
+    it('should not throw for an empty string', () => {
+      testEstimation('', 0)
+    })
+
+    it('should compute the number of tokens in an array of integer inputs', () => {
+      testEstimation([1, 2, 3], 3)
+    })
+
+    it('should compute no tokens for invalid content', () => {
+      testEstimation({}, 0)
+    })
+  })
+})
